2020 October update
New:
  * grc
  * rtv
  * swappy
  * clight
  * script for starting f5pvn (f5vpn.sh)
  * script for fixing blueooth connection (fix-bluetooth)
  * script for sway idle (idle.sh)
  * umount in backup script
  * italics terminfo

Changes:
  * alacritty color config
  * sway config
  * yay configuration: clean, don't remove make deps
  * git diff, graph
  * mutt mailboxes, colors

+++ bin/mutt-open.sh
@@ -1,9 +1,16 @@
-#!/usr/bin/env sh
+#!/usr/bin/env bash
 set -euo pipefail
 
-echo $@
-TMP=$(basename $1)
-TMP=~/.cache/mutt.tmp."${TMP##*.}"
-TMP="${TMP%%__}"
-cp -f "$1" "$TMP"
-xdg-open "$TMP" 2>/dev/null
+echo "$@"
+
+if [ -f "$1" ]; then
+  TMP=$(basename "$1")
+  TMP=~/.cache/mutt.tmp."${TMP##*.}"
+  TMP="${TMP%%__}"
+
+  cp -f "$1" "$TMP"
+  xdg-open "$TMP" 2>/dev/null
+else
+  xdg-open "$1" 2>/dev/null
+fi
+
